pyproject.toml
[tool.poetry]
name = "robotframework-gpio"
version = "0.1.0-beta"
description = "Robot Framework Library for interfacing GPIO pins on executing robot files on Raspberry Pi's. "
authors = ["Yusuf Can Bayrak <yusufcanbayrak@gmail.com>"]
license = "Apache-2.0"
readme = 'README.md'
packages = [
    { include = "src/GPIOLibrary" },
    { include = "utests", format = "sdist" }
]

[tool.poetry.dependencies]
python = "^3.6"
robotframework = "^3.2.2"

[tool.poetry.dev-dependencies]
coverage = "^5.4"
pytest = "^6.2.2"
pytest-coverage = "^0.0"
tox = "^3.22.0"
black = "^20.8b1"
twine = "^3.3.0"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "utests",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
